requirements.txt
pdf2docx
python-docx
reportlab
pypandoc